requirements.txt
snowflake-snowpark-python[pandas]>=1.7.0
streamlit==1.28.1
#openai==1.2.2
google-cloud-aiplatform>=1.38
vertexai>=0.0.1
google-generativeai>=0.3.1
#google-api-python-client
#pytest
#from google.cloud import aiplatform
#from vertexai.generative_models import GenerativeModel, ChatSession